requirements.txt
# Multi-Agent Tutorial Dependencies
# Python 3.10+ required

# ============================================
# Core LLM & Agent Frameworks
# ============================================

# LangChain ecosystem
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0

# Microsoft AutoGen
pyautogen>=0.4.0

# CrewAI (optional, for comparison)
crewai>=0.80.0

# ============================================
# LLM Providers
# ============================================

openai>=1.50.0
anthropic>=0.37.0
google-generativeai>=0.8.0

# Local LLM support
ollama>=0.3.0

# ============================================
# Agent Tools & Utilities
# ============================================

# Code execution
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Web & API
requests>=2.32.0
httpx>=0.27.0
aiohttp>=3.10.0

# File operations
aiofiles>=24.1.0

# ============================================
# Docker & Sandboxing
# ============================================

docker>=7.1.0

# ============================================
# Evaluation & Testing
# ============================================

# DeepEval for agent evaluation
deepeval>=1.4.0

# Ragas for RAG evaluation
ragas>=0.2.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# ============================================
# Observability & Monitoring
# ============================================

# Arize Phoenix
arize-phoenix>=5.0.0

# OpenTelemetry
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0

# ============================================
# Data Processing
# ============================================

pandas>=2.2.0
numpy>=2.1.0

# ============================================
# Utilities
# ============================================

# Rich console output
rich>=13.9.0

# Progress bars
tqdm>=4.66.0

# YAML processing
pyyaml>=6.0.0

# JSON handling
orjson>=3.10.0

# Async utilities
tenacity>=9.0.0

# ============================================
# Development Tools (optional)
# ============================================

# Code formatting
black>=24.10.0
isort>=5.13.0

# Linting
ruff>=0.7.0
mypy>=1.13.0

# Jupyter support
jupyter>=1.1.0
ipykernel>=6.29.0
